Overhaul main guide page and add controller pages
The main guide page is at the moment pretty minimal, and assumes a fair
amount of knowledge on the part of the user. This commit aims to
overhaul it, providing all the information a beginner needs to setup the
emulator, while still being quick to read. The main page now has images
and more detailed explanations for settings, has warnings about common
pitfalls, and now uses content tabs to organize information.

Additionally the commit adds new controller-specific setup guide pages,
along with the profiles needed for xbox. gitattributes was changed to
stop git from changing the line endings in the profile files.

I've also added an FAQ with frequent questions, that can hopefully be
expanded.

These changes are bundled as a single commit because they reference
eachother and new files, and can't really be applied atomically.